feat(cmd/rofl): Add basic proxy support
